Create a test containing a join-based deadlock
Adding edges between the threads that join each other in the waits-for graph is basically how one can detect deadlocks caused by mutual joins. The edges are added in the same waits-for graph as the one used for mutex-based deadlocks. It should be mentioned that the edges are not removed during and after the join operation, particularly because removing the edges would lead to a failed detection.